Modulo de areas, impresion de constancias

+++ creditsch/app/Http/Controllers/CreditosController.php
@@ -4,8 +4,12 @@ namespace App\Http\Controllers;
 
 use Illuminate\Http\Request;
 use App\Credito; //Es el nombre del modelo con el que va a trabajar el controlador
+use App\Area;
+use App\CreditoArea;
+use App\User;
 use Laracasts\Flash\Flash; //Es el paquete para poder usar los mensajes de alerta tipo bootstrap
 use Spatie\Permission\Models\Permission;
+use DB;
 
 class CreditosController extends Controller
 {
@@ -26,6 +30,7 @@ class CreditosController extends Controller
     {
         //Aqui mandamos llamar la vista de la pagina de inicio de alumnos
         $credito=Credito::orderby('id','asc')->paginate(5); //Consulta todos los usuarios y los ordena, ademas pagina la consulta
+        $credito = Credito::leftjoin('users as u','u.id','=','creditos.credito_jefe')->select('creditos.nombre','creditos.vigente','creditos.id','u.name as jefe_nombre')->paginate(5);
         return view('admin.creditos.index')->with('credito',$credito); //Llama a la vista y le envia los usuarios
     }
 
@@ -37,7 +42,11 @@ class CreditosController extends Controller
     public function create()
     {
         //Ponemos el codigo de la vista que se llamara para las altas de los creditos
-        return view('admin.creditos.create');
+        $areas = Area::orderby('nombre','ASC')->get();
+        $usuarios = User::where('active','=','true')->get();
+        return view('admin.creditos.create')
+        ->with('areas',$areas)
+        ->with('usuarios',$usuarios);
     }
 
     /**
@@ -53,6 +62,13 @@ class CreditosController extends Controller
         $credito = new Credito($request->all());
         //Comando para guardar el registro
         $credito->save();
+        for($x=0; $x<count($request->areas); $x++){
+            $credito_area = new CreditoArea();
+            $credito_area->timestamps = false;
+            $credito_area->credito_id=$credito->id;
+            $credito_area->credito_area=$request->areas[$x];
+            $credito_area->save();
+        }
         Flash::success('El credito  '.$credito->nombre.' se ha registrado de forma exitosa');
         return redirect()->route('creditos.index');
     }
@@ -77,8 +93,20 @@ class CreditosController extends Controller
     public function edit($id)
     {
         //Codigo de modificaciones
+        $areas = DB::table('areas as a')->leftjoin('creditos_areas as ca', function($join) use($id){
+            $join->on('ca.credito_area','=','a.id');
+            $join->where('ca.credito_id','=',$id);
+        })->select('a.nombre','a.id','ca.credito_id')->orderBy('nombre','ASC')->get();
         $credito=Credito::find($id);//Busca el registro
-        return view('admin.creditos.edit')->with('credito',$credito);
+        if($credito==null){
+            Flash::error('El credito no existe');
+            return redirect()->route('creditos.index');
+        }
+        $usuarios = User::where('active','=','true')->orwhere('id','=',$credito->credito_jefe)->get();
+        return view('admin.creditos.edit')
+        ->with('credito',$credito)
+        ->with('areas',$areas)
+        ->with('usuarios',$usuarios);
     }
 
     /**
@@ -90,10 +118,20 @@ class CreditosController extends Controller
      */
     public function update(Request $request, $id)
     {
+        CreditoArea::where('credito_id','=',$id)->delete();
         //Ejecuta la modificacion
         $credito= Credito::find($id);
         $credito->nombre=$request->nombre;
+        $credito->credito_jefe = $request->credito_jefe;
+        $credito->vigente = $request->vigente;
         $credito->save();
+        for($x=0; $x<count($request->areas); $x++){
+            $credito_area = new CreditoArea();
+            $credito_area->timestamps = false;
+            $credito_area->credito_id=$id;
+            $credito_area->credito_area=$request->areas[$x];
+            $credito_area->save();
+        }
         Flash::warning('El credito '. $credito->nombre .' a sido editado de forma exitosa');//Envia mensaje
         return redirect('admin/creditos');//llama a la pagina de consultas
     }
@@ -108,6 +146,23 @@ class CreditosController extends Controller
     {
         //Codigo de bajas
         $credito=Credito::find($id);//Busca el registro
+        if($credito==null){
+            Flash::error('El credito no existe');
+            return redirect()->route('creditos.index');
+        }
+        $tiene_actividades = DB::table('actividad as a')->join('creditos as c',function($join) use($id){
+            $join->on('a.id_actividad','=','c.id');
+            $join->where('c.id','=',$id);
+        })->select('c.id')->get()->count()>0?true:false;
+
+        $tiene_avance = DB::table('avance as a')->join('creditos as c', function($join) use($id){
+            $join->on('c.id','=','a.id_credito');
+            $join->where('c.id','=',$id);
+        })->select('c.id')->get()->count()>0?true:false;
+        if($tiene_avance || $tiene_actividades){
+            Flash::error('No se puede eliminar debido a claves foraneas');
+            return redirect()->route('creditos.index');
+        }
         $credito->delete();//Elimina el registro
         Flash::error('El credito '. $credito->nombre .' a sido borrado de forma exitosa');//Envia mensaje
         return redirect('admin/creditos');//llama a la pagina de consultas
