+++ creditsch/app/Http/Controllers/UsersController.php
@@ -6,6 +6,7 @@ use App\Http\Requests\UserRequest;
 use Illuminate\Http\Request;
 use Illuminate\Support\Facades\Auth;
 use App\User;
+use App\Area;
 use Laracasts\Flash\Flash;
 use Illuminate\Support\Facades\Hash;
 use Spatie\Permission\Models\Role;
@@ -22,21 +23,28 @@ class UsersController extends Controller
         $this->middleware('permission:VIP|ASIGNAR_REMOVER_ROLES_USUARIOS')->only(['guardarRoles','asignarRoles']);
     }
     public function index(){
-    	$users = User::all();
-    	return view('admin.usuarios.index')
-    	->with('users',$users);
+        if (Auth::User()->can('VIP')) {
+                $users = DB::table('users as u')->join('areas as a','a.id','=','u.area')->select('a.nombre as area','u.name','u.active','u.id','u.email')->get();
+                return view('admin.usuarios.index')
+                ->with('users',$users);
+        }else{
+                $users = DB::table('users as u')->join('areas as a','a.id','=','u.area')->where([
+                    ['u.area','=',Auth::User()->area],
+                    ['u.id','<>','1']
+                ])->select('a.nombre as area','u.name','u.active','u.id','u.email')->get();
+                return view('admin.usuarios.index')
+                ->with('users',$users);
+        }
+        
     }
     public function create(){
-    	$carreras = [
-    		['carrera' => 'Ing. en Sistemas Computacionales','valor' => 'Sistemas'],
-    		['carrera' => 'Ing. en Nanotecnología', 'valor' => 'Nanotecnología'],
-    		['carrera' => 'Ing. en Mecatrónica','valor' => 'Mecatrónica'],
-    		['carrera' => 'Ing. en Bioquímica','valor' => 'Bioquímica'],
-    		['carrera' => 'Ing. en Tecnologías de la Información y Comunicaciones','valor' =>"TIC's"],
-    		['carrera' => 'Ing. en Gestión Empresarial','valor' => 'Gestión Empresarial'],
-    		['carrera' => 'Ing. Industrial', 'valor' => 'Industrial']
-    	];
-    	return view('admin.usuarios.registrar')->with('carreras',$carreras);
+        if (Auth::User()->can('VIP')) {
+            $areas = Area::orderBy('nombre','ASC')->get();
+            return view('admin.usuarios.registrar')->with('areas',$areas);
+        }else{
+            $areas = Area::where('id','=',Auth::User()->area)->orderBy('nombre','ASC')->get();
+            return view('admin.usuarios.registrar')->with('areas',$areas);
+        }
     }
     public function store(UserRequest $request){
     	$user = new User($request->all());
@@ -47,23 +55,39 @@ class UsersController extends Controller
     }
 
     public function edit($id){
-        $carreras = [
-            ['carrera' => 'Ing. en Sistemas Computacionales','valor' => 'Sistemas'],
-            ['carrera' => 'Ing. en Nanotecnología', 'valor' => 'Nanotecnología'],
-            ['carrera' => 'Ing. en Mecatrónica','valor' => 'Mecatrónica'],
-            ['carrera' => 'Ing. en Bioquímica','valor' => 'Bioquímica'],
-            ['carrera' => 'Ing. en Tecnologías de la Información y Comunicaciones','valor' =>"TIC's"],
-            ['carrera' => 'Ing. en Gestión Empresarial','valor' => 'Gestión Empresarial'],
-            ['carrera' => 'Ing. Industrial', 'valor' => 'Industrial']
-        ];
         $user = User::find($id);
-        return view('admin.usuarios.edit')
-        ->with('carreras',$carreras)
-        ->with('user',$user);
+        if($user==null){
+            Flash::error('El usuario no existe');
+            return redirect()->back();
+        }
+        if (Auth::User()->can('VIP')) {
+            $areas = Area::orderBy('nombre','ASC')->get();
+            return view('admin.usuarios.edit')
+            ->with('areas',$areas)
+            ->with('user',$user);
+        }else{
+            if($user->area!=Auth::User()->area){
+                Flash::error('No puedes editar usuarios que no te corresponden');
+                return redirect()->back();
+            }
+            if($user->id==1){
+                Flash::error('No le puedes hacer nada al administrador');
+                return redirect()->back();
+            }
+            $areas = Area::where('id','=',Auth::User()->area)->orderBy('nombre','ASC')->get();
+            return view('admin.usuarios.edit')
+            ->with('areas',$areas)
+            ->with('user',$user);
+        }
+        
     }
 
     public function update(UserRequest $request, $id){
         $user = User::find($id);
+        if($user==null){
+            Flash::error('El usuario no existe');
+            return redirect()->back();
+        }
         $user->name = $request->name;
         $user->email = $request->email;
         $user->area = $request->area;
@@ -83,16 +107,39 @@ class UsersController extends Controller
             Flash::error('El usuario no existe');
             return redirect()->route('usuarios.index');
         }
-        $actividades = DB::table('actividad as a')->where('a.id_user','=',$id)->get()->count()>0? true: false;
-        $responsable = DB::table('actividad_evidencia as ae')->where('ae.user_id','=',$id)->orwhere('ae.user_id','=',$id)->get()->count()>0?true:false;
-        $roles = DB::table('model_has_roles')->where('model_id','=',$id)->get()->count()>0?true:false;
 
-        if($roles || $actividades || $responsable){
-            Flash::error('El usuarios '.$user->name.' no puede ser eliminado debido debido a claves foraneas');
-            return redirect()->route('usuarios.index');
+        if(Auth::User()->can('VIP')){
+            $actividades = DB::table('actividad as a')->where('a.id_user','=',$id)->get()->count()>0? true: false;
+            $responsable = DB::table('actividad_evidencia as ae')->where('ae.user_id','=',$id)->orwhere('ae.user_id','=',$id)->get()->count()>0?true:false;
+            $roles = DB::table('model_has_roles')->where('model_id','=',$id)->get()->count()>0?true:false;
+
+            if($roles || $actividades || $responsable){
+                Flash::error('El usuarios '.$user->name.' no puede ser eliminado debido debido a claves foraneas');
+                return redirect()->route('usuarios.index');
+            }
+            $user->delete();
+            return redirect()->back();
+        }else{
+            if($user->area!=Auth::User()->area){
+                Flash::error('No puedes eliminar usuarios que no te corresponden');
+                return redirect()->back();
+            }
+            if($user->id==1){
+                Flash::error('No le puedes hacer nada al administrador');
+                return redirect()->back();
+            }
+            $actividades = DB::table('actividad as a')->where('a.id_user','=',$id)->get()->count()>0? true: false;
+            $responsable = DB::table('actividad_evidencia as ae')->where('ae.user_id','=',$id)->orwhere('ae.user_id','=',$id)->get()->count()>0?true:false;
+            $roles = DB::table('model_has_roles')->where('model_id','=',$id)->get()->count()>0?true:false;
+
+            if($roles || $actividades || $responsable){
+                Flash::error('El usuarios '.$user->name.' no puede ser eliminado debido debido a claves foraneas');
+                return redirect()->route('usuarios.index');
+            }
+            $user->delete();
+            return redirect()->back();
         }
-        $user->delete();
-        return redirect()->back();
+        
     }
 
     public function asignarRoles($id){
